requirements.txt
typer
Pillow
pillow-heif
rich